requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
certifi==2024.6.2
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.4
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.3
pydantic-core==2.18.4
pydantic-extra-types==2.8.0
pydantic-settings==2.3.2
pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
